+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20201109133210.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Remove deprecated "scanned" field from "contact" table
+ */
+final class Version20201109133210 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE contact DROP scanned');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE contact ADD scanned TINYINT(1) NOT NULL');
+        $this->addSql('UPDATE contact SET contact.scanned = IF(contact.origin = \'scan\', 1, 0)');
+    }
+}